Configure the proxy in PESigner and use system proxy if nothing else was set
